ulux-binding: Reorganize message handler package structure
