+++ bundles/binding/org.openhab.binding.ulux.test/src/test/java/org/openhab/binding/ulux/internal/handler/messages/LuxMessageHandlerTest.java
@@ -1,4 +1,4 @@
-package org.openhab.binding.ulux.internal.handler;
+package org.openhab.binding.ulux.internal.handler.messages;
 
 import static org.mockito.Mockito.verify;
