+++ bundles/binding/org.openhab.binding.ulux/src/main/java/org/openhab/binding/ulux/internal/handler/UluxMessageHandlerFacade.java
@@ -26,6 +26,19 @@ import java.util.EnumMap;
 import java.util.Map;
 
 import org.openhab.binding.ulux.UluxBindingProvider;
+import org.openhab.binding.ulux.internal.handler.messages.AbstractMessageHandler;
+import org.openhab.binding.ulux.internal.handler.messages.AudioVolumeMessageHandler;
+import org.openhab.binding.ulux.internal.handler.messages.ControlMessageHandler;
+import org.openhab.binding.ulux.internal.handler.messages.EditValueMessageHandler;
+import org.openhab.binding.ulux.internal.handler.messages.EventMessageHandler;
+import org.openhab.binding.ulux.internal.handler.messages.IdListMessageHandler;
+import org.openhab.binding.ulux.internal.handler.messages.LedMessageHandler;
+import org.openhab.binding.ulux.internal.handler.messages.LuxMessageHandler;
+import org.openhab.binding.ulux.internal.handler.messages.PageCountMessageHandler;
+import org.openhab.binding.ulux.internal.handler.messages.PageIndexMessageHandler;
+import org.openhab.binding.ulux.internal.handler.messages.StateMessageHandler;
+import org.openhab.binding.ulux.internal.handler.messages.UluxMessageHandler;
+import org.openhab.binding.ulux.internal.handler.messages.VideoStateMessageHandler;
 import org.openhab.binding.ulux.internal.ump.UluxMessage;
 import org.openhab.binding.ulux.internal.ump.UluxMessageDatagram;
 import org.openhab.binding.ulux.internal.ump.UluxMessageId;
