+++ bundles/binding/org.openhab.binding.ulux/src/main/java/org/openhab/binding/ulux/internal/handler/messages/AudioVolumeMessageHandler.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.ulux.internal.handler;
+package org.openhab.binding.ulux.internal.handler.messages;
 
 import static org.openhab.binding.ulux.UluxBindingConfigType.AUDIO_VOLUME;
 
@@ -21,7 +21,7 @@ import org.openhab.core.library.types.PercentType;
  * @author Andreas Brenk
  * @since 1.7.0
  */
-final class AudioVolumeMessageHandler extends AbstractMessageHandler<AudioVolumeMessage> {
+public final class AudioVolumeMessageHandler extends AbstractMessageHandler<AudioVolumeMessage> {
 
 	@Override
 	public void handleMessage(AudioVolumeMessage message, UluxMessageDatagram response) {
