+++ bundles/binding/org.openhab.binding.ulux/src/main/java/org/openhab/binding/ulux/internal/handler/messages/LedMessageHandler.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.ulux.internal.handler;
+package org.openhab.binding.ulux.internal.handler.messages;
 
 import static org.openhab.binding.ulux.UluxBindingConfigType.LED;
 
@@ -22,7 +22,7 @@ import org.openhab.core.library.types.DecimalType;
  * @author Andreas Brenk
  * @since 1.7.0
  */
-final class LedMessageHandler extends AbstractMessageHandler<LedMessage> {
+public final class LedMessageHandler extends AbstractMessageHandler<LedMessage> {
 
 	@Override
 	public void handleMessage(LedMessage message, UluxMessageDatagram response) {
