+++ bundles/binding/org.openhab.binding.ulux/src/main/java/org/openhab/binding/ulux/internal/handler/messages/StateMessageHandler.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.ulux.internal.handler;
+package org.openhab.binding.ulux.internal.handler.messages;
 
 import static org.openhab.binding.ulux.UluxBindingConfigType.AMBIENT_LIGHT;
 import static org.openhab.binding.ulux.UluxBindingConfigType.AUDIO;
@@ -26,7 +26,7 @@ import org.openhab.core.library.types.OnOffType;
  * @author Andreas Brenk
  * @since 1.7.0
  */
-final class StateMessageHandler extends AbstractMessageHandler<StateMessage> {
+public final class StateMessageHandler extends AbstractMessageHandler<StateMessage> {
 
 	@Override
 	public void handleMessage(StateMessage message, UluxMessageDatagram response) {
